requirements.txt
# Mangapill Downloader Requirements

# CLI Framework
typer[all]>=0.9.0

# Rich Terminal Formatting
rich>=13.0.0

# HTTP Requests
requests>=2.31.0

# HTML Parsing
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Image Processing
Pillow>=10.0.0

# PDF Generation
img2pdf>=0.5.0
